cleanup inTree()
It used to be the case that the presence (non-NULL-ness)
of parentSymbol was not enough to decide Expr::inTree().

This PR
-------

* establishes Expr::inTree() === (Expr::parentSymbol != NULL)

* adjusts Expr::inTree() correspondingly,

* replaces a few explicit checks for "parentSymbol != NULL"
  with calls to inTree(), for clarity.

The only case where this discipline did not hold was for ModuleSymbol::block.
I adjusted that, almost trivially.

While there, I made "this == rootModule" the second check in Symbol::inTree().
Because we expect the now-first check to succeed in most cases and
the "this == rootModule" check to fail almost always.

While there, I made {Expr,Type,Symbol}::inTree() non-virtual,
as we do not override them in subclasses.

Other potential improvements
----------------------------

* There may be other cases where we could replace "parentSymbol != NULL"
with inTree().

* Add INT_ASSERT(symbol != NULL) in Type::verify().

* Replace isAlive() with inTree().

The only reason not to do that that I see is that isAlive() executes
slightly fewer conditionals than inTree(). This could be mitigated:

  - To avoid the check for rootModule, we can add another version of
    Symbol::inTree(). We could apply that version whenever we do not need the
    precise answer for rootModule, ex. when the Symbol is not a ModuleSymbol.
    Maybe Symbol::inTree() could be virtual, overriden for ModuleSymbol.

  - To avoid the check for Type::symbol!=NULL, we could ensure that
    it is always the case. Add INT_ASSERT(symbol!=NULL) to Type::verify().