Add a scalable AllLocalesBarrier
This adds a barrier that must be called on all locales, optionally with a
certain number of tasks per locale.

The implementation is a little ugly, because it hacks into the existing barrier
code (to avoid code duplication) and calls a comm_layer_barrier that only
supports a single barrier (i.e. it's a global barrier for some comm layers, so
only one AllLocalesBarrier can be created)