Adjust addArgCoercion to skip deref before de-sync
Say a ref(sync(int)) is passed to an int argument. Before
this commit, the compiler would perform the following coercions:
 ref(sync(int)) -> sync(int) -> int

Note that the sync(int) -> int call is achieved with readFE().

But, for some time now, it has been legal to provide a 'ref'
as an argument for a 'ref'-intent function call. Thus,
the middle conversion is no longer necessary and we can now
do the conversion in 1 step by calling readFE() and passing
in the ref(sync(int)) actual.