Achieve single-node execution.
It turned out we only needed two things to reach single-node execution.
First, since fi_freeinfo() uses free() on the provider name member of
the fabric attributes (this is documented, though obscurely), we needed
to either allocate that from the system (malloc() e.g.) or free it and
NULL the pointer before calling fi_freeinfo().  Here, choose the former
and acquire the space via sys_malloc().

Also, fix a race which allowed us to start tearing everything down (in
chpl_comm_pre_task_exit()) before we had even finished setting it up.
For this we use a pthreads condition variable and associated mutex, in
both the setup and the teardown code.

While here, clean up the non-Cray-XC out-of-band code somewhat, such
that for now we only support (and only minimally for now) a solution
based on sockets and only under slurm.

With these changes in place a single-node hello-world execution works.
A 2-node one still fails, though, with gripes about unimplemented GET
and PUT on Cray XC and unimplemented out-of-band barrier elsewhere.