coforall/try! bug (#8451)
[reviewed by @mppf]

@benharsh discovered a bug in error handling where a `try!` at the end of a `coforall` block allows unhandled throwing calls to escape within the `coforall`. 

The error ended up being in `ImplicitThrowsVisitor`, where `exitTryStmt()` would mark an entire function block as non-throwing if `try!` was the last statement. This was triggered by a `coforall` because they are lowered into a `coforall_fn()` during compilation.

This change required additional checking on `throw` statements in order to maintain correctness.