Adjust needToAddCoercion to avoid deref for new in
addArgCoercion was adding a deref, and that resulted
in the generated code trying to auto-destroy the
array twice.

This commit adjusts needToAddCoercion to avoid adding
a coercion in that case, since it's not necessary when
the next thing we will do is pass it to initCopy.

See the test

 functions/vass/setter-passed-to-another-fun

In particular, this pattern:

  var globalArr: [1..2] int;
  proc accessorArr() ref {
    writeln(true);
    return globalArr;
  }
  proc accessorArr() {
    writeln(false);
    return globalArr;
  }

  f_in(accessorArr());