Add another future relating to returned array size checking
With a bit more experimentation, it turns out that the behavior for
returning arrays of different size than the declared return type is
very different for promoted expressions vs. not (I believe that's the
differentiator).  When returning the promoted expression as in
checkReturnedArray.chpl, we truncate the expression to the declared
return type (or get an OOB error if the returned domain is larger than
the expression).  Whereas with other expressions, we just ignore the
domain of the returned array altogether (so it can be a lie and
nothing will complain).

Here, I'm filling a test for the second case while updating the
future of the original to note this new observation.