Put the main process to sleep while waiting to shutdown
This PR frees up `chpl_comm_barrier()` for use in module code, and prevents the
main process from interfering with tasks running on the same CPU.

Previously, the main process on non-0 locales would spin-wait on a barrier in
`chpl_comm_pre_task_exit()`, waiting for locale-0 to tell them to do an orderly
shutdown. This barrier prevented chpl_comm_barrier from being used in user
applications, and the spin-waiting hurts performance for tasks running on the
same CPU. This alleviates both issues by putting the main process to sleep with
a condition variable, and having locale-0 wake everybody up with a AM that
signals the cond var when it's time to shutdown.

This is motivated by an upcoming scalable barrier PR that wants to use
`chpl_comm_barrier()` in module code. The scalable barrier will significantly
improve ISx performance and scalability. Putting the main process to sleep
reduces interference, but we probably won't notice this in any tests yet
because the progress-thread currently causes more interference. i.e. tasks
sharing a core with the progress-thread are already the slowest tasks, so the
benefit of parking the main process will only be visible once that issue is
resolved.