Adjust early lvalue checking to allow Owned temps
Early lvalue checking was causing problems with several
patterns with Owned because it's easy to write Owned use
cases where a call-temp needs to be passable to a Ref argument.

This commit adjusts early lvalue checking in resolution to
allow this pattern and additionally changes findNonTaskFnParent
to avoid asserting for the pattern that happens to occur with Owned.