Add low-level communication performance testing.
This adds perf testing for low-level communication: GETs, PUTs,
on-stmts, and "fast" on-stmts.  So far the only test is for many-to-one
communication, where N-1 nodes all communicate with the same target
node.  We envision adding one-to-one and one-to-many testing also.  We
may add many-to-many testing, although the results of that may tell us
more about the network than about Chapel communication, so it's
secondary.  So far the ml-compopts are only set up for comm=ugni.