Correct ISx reported times
Previously, if numTrials was greater than 1 our timing values weren't correct.
Instead of computing the average across all trials and tasks, we took the min
for each task from any trial. e.g. if we had timings like:

|         | task 1 | task 2 |
| ------- | ------ | ------ |
| trial 1 |   5s   |     1s |
| trial 2 |   1s   |     5s |

We'd report an average time of 1s per task, which looks great but isn't
accurate. We will now correctly report the average for each task as 3 seconds.
Note that this isn't an issue for nightly testing where we only do 1 trial, but
this was giving misleading results for scalability numbers where we tend to run
5 trials.