Add a barrier performance test
This adds a barrier micro-benchmark to track the performance of barriers for
both single and multi-locale. It's somewhat lame, in the sense that it's
testing a barrier by repeatedly doing barrier.barrier() in a trial loop, so all
tasks will be perfectly aligned/synchronize, which doesn't happen in real
applications. That said, it's better than nothing and will help track upcoming
improvements to barriers used across multiple locales. This only tests the
atomic barrier, since the sync one isn't reusable.

Our barrier performance is terrible for multi-locale (we knew/expected this)
since all locales have to go back to the locale that owns the barrier. This
poor scalability significantly hurts ISx. We have plans to add a more scalable
barrier (which is part of the motivation for adding this test.)

As a side node, our atomic barrier doesn't scale terribly well for a single
node either. We use a pretty simple linear barrier (centralized atomic counter
and wait variables.) This creates a lot of contention and cache-thrash. We
probably want some sort of dissemination barrier even within a single node.
There's a ton of papers/research out there on the best barrier for a given
situation (it's highly dependent on core/node count) so I'm sure we can lean on
those whenever we end up needing a better barrier.

The basic plan for the more scalable multi-locale barrier is to do a local
barrier on each locale, and last task to enter will do a chpl_comm_barrier().
For ISx, the poor scalability of the local barrier won't really hurt us, since
ISx only does 1 or 2 barriers in the timed section.