bug: "alist must be indexed by positive integer (generic inheritance + inits)
This test hits an internal error we don't see often, possibly due to
the lack of fields in the classes.  It doesn't seem to matter whether
I use the compiler-generated initializer or introduce one myself
(though I've only done the former here, for maintenance reasons...  I
think we ultimately want the version with a compiler initializer to
work and this will help us find it when we remove the pragma).