Test showing initializer problems when chaining dmaps, doms, arrs
This test shows that when chaining dmaps, domains, and arrays problems
occur.  For a --local compilation, it seems as though things just
aren't set up correctly, which I attributed to domain maps not yet
using initializers, but which could actually be due to issue #8503.

Weirdly, though, when run on multiple locales, the test seems to work
as expected until the end when a derefernce of nil is hit.  I only hit
this late in filing the bug and haven't had a chance to look into it
yet.