Link in libmvec.a when necessary (#8432)
Link in libmvec.a when necessary (#8432)

[reviewed by @ronawho and @gbtitus ]

Gcc now needs a vectorized math library, libmvec. This is made available automatically in most cases, but when linking statically on Linux, it needs to be specified explicitly on the link command line.

When using gcc, this change looks to see if libmvec is available, and if so, assumes it is needed whenever we are linking statically.

The -lmvec flag for the vectorized math library must appear before the -lm flag for the regular math library, according to https://sourceware.org/glibc/wiki/libmvec .

Tested on Ubuntu, SLES, Mac, and Intel/AMD/ARM-based Cray systems.