Add support for casting arrays to strings
It seems that, previously, casting arrays to strings was not supported
directly and would sometimes fall into a promotion case before falling
into the catch-all 'stringify'-based rematch that occurs at the bottom
of ChapelIO.chpl.  This resulted in seemingly inconsistent results
from case to case depending on the array's type.  This PR attempts to
introduce a general case for casting arrays to strings and adds in
the tests that showed inconsistencies before (from issue #8513).