Set rootLocaleInitialized in RootLocale.deinit
Before this commit, RootLocale deinit consisted of a
for loop over the locales, where it would delete each
locale in turn. Note though that deleteing an object
runs an 'on' statement to go where the object is stored.
After PR #7193, these 'on' statements had the effect
of changing the running task count field of the locale
object - which in some cases was already deleted.

This commit sets rootLocaleInitialized=false in this
loop in order to disable the modifications to the
running task count. It wraps both this change an the
delete in an on statement.