Add support for -futures-only to paratest.server
start_test has long supported a -futures-only flag that tested the
future tests but no others.  paratest.server (and by extension,
paratest.chapcs) has _seemed_ to support such a flag, in that it
accepted it; but in practice, it ran all tests and futures because
the flag parsing for paratest.server is weak and checks for prefix
strings rather than complete strings.  So it fell into the -futures
case silently and ran everything.

Being able to run paratest in this mode has long seemed attractive
to me because I commonly run paratest.chapcs, get clean output,
and then think "Oh shoot, I should've checked futures as well"
so re-run the whole thing, wasting time and cycles.

Here, I added support for a -futures-only flag, building on
pre-existing support in subsequent scripts for this being supported
via futures mode "2".  Thanks to Elliot for doing the detective
work to determine that it would be this easy to implement.

Note that my futures-only flag is as lame as the other paratest
flags in that you could also say -futures-only-stupid-script
and it would accept the flag and run the futures only.