Add an array-of-barriers test that fails on master
Default initialization of an array of barriers fails to resolve because we
don't have a 0-arg initializer. This is a small reproducer for a scalable
AllLocales barrier that I'm working on.

Beyond an array-of-barriers, just declaring a type as a barrier without
initializing it leads to a resolution error:

```chpl
var bar: Barrier;
```