Fix problem with serialization+new in intent
After PR #8417, the pattern of where destroys should
be called for in intent changed. In particular, an in
intent argument now transfers responsibility to destroy
a value from the call site to the called function.

This commit adjusts insertSerialization to account for the
case that the 'on' function already includes a destruction
call for an 'in' intent argument. In that event, there is
no need to add an additional destruction call.

Resolves failures with
  parallel/forall/in-intents/both-arr-dom-const-var
  parallel/forall/in-intents/domain-var-var

in CHPL_LOCALE_MODEL=numa testing.