Add a test of a failure mode with default initializers for classes
When a class with a default initializer is created by relying on the default
value for it record field (which has an initializer), we encounter an error
due to not creating the default value correctly.  This is not a problem for
explicit initializers having the same argument declaration, nor for default
constructors.