I missed a flag (or maybe the flag got added?) for default initializers
This flag (FLAG_SUPPRESS_LVALUE_ERRORS) was necessary for the default value of
the record field corresponding argument when record fields had initializers of
any form.  Otherwise, we would get errors about improperly using lvalues.  This
change adds the flag to the default initializer, and propagates it to the _new
function if the type is a class.  It should resolve ~27 failures with
--force-initializers