Speed up ugni's barrier by using chained transactions
Use a vector/chained put to wake up the child locales for the ugni barrier.
This increases the number of children from 2 to MAX_CHAINED_PUT_LEN, which
reduces the depth of the tree used to barrier. Since each level in the tree
adds ~1 network trip, reducing the depth helps improve performance. Below are
some perf results with and without this PR for 1,000,000 trials of the
following barrier micro-benchmark:

    coforall loc in Locales do on loc do
      for 1..numTrials do
          chpl_comm_barrier(msg);

| Barrier Type (1M trials)  | 2 node | 4 node | 16 node | 64 node | 256 node |
| ------------------------- | ------ | ------ | ------- | ------- |--------- |
| master chpl_comm_barrier  |  3.4s  |  5.1s  | 13.4s   | 24.2s   | 37.6s    |
| chained chpl_comm_barrier |  3.4s  |  4.8s  |  9.8s   | 20.2s   | 29.9s    |

I don't expect this to have much of an impact for any current benchmarks
(beyond the barrier micro-benchmark.) ISx is the most likely benchmark to be
impacted, but it doesn't do enough barriers for this change to have noticeable
impact.