Specify how long to run rather than how many ops to do.
This makes a wholesale change in how the program's effort is sized.  We
used to specify the number of ops to do, with the compopts files picking
values based on the particular op and the configured comm layer.  Switch
to specifying the time to run instead.  We do a some number of initial
ops to compute a tentative rate, and then work our way up to the target
time through incremental estimates of the number of ops needed to get us
there.  We may run a little long but won't ever run short.