Add testing for a fetching AMO operator.
We could only test non-fetching AMO before.  Add a fetching AMO op.  The
difference is that the fetching AMO might have extra work on the remote
side to send back the result.  (In practice on Cray XC/Aries systems
there does not seem to be much difference, unlike PUT vs. GET.  This may
well be because the target datum is just residing in the NIC AMO cache
for the entire duration of the test.)

While here, change the AMO op to be an add rather than a xor, just
because it seems more "basic" to me.