Fix error with enum casts for enums w/ nontrivial int expressions
This fixes a longstanding bug noted in issue #5416 in which the cast
functions that we built for enums didn't work properly if the
expression used to define the enum was not a simple integer literal.
The old logic took the approach that if the integer expression
defining the enum was not a literal, it might as well not be there and
incremented its counter of what it thought the global value was.
Here, I implement a new approach where I keep track of the last
expression found and the number of fields since that last expression
was found and create a conditional that adds those two values
together.  This seems more correct, though more testing is probably in
order.