Add support for multiple registries to Mason
Update Mason to support multiple registries.  The environment variable
`MASON_REGISTRY` is used to list the available registries. It is a
comma-separated list of "name:location" pairs where "name" is what
the registry should be named locally and location is either a local
path or a URL pointing to a registry.  If unset, it defaults to
"registry:https://github.com/chapel-lang/mason-registry", which
matches the behavior before this change.

If `MASON_REGISTRY` includes multiple registries, they are searched in
left-to-right order for any dependencies.

If `MASON_CACHED_REGISTRY` is set, it should be a comma-separated list
of directories the same length as the number of "name:location" pairs
in `MASON_REGISTRY`.