Modify the handling of new expressions for argument default values
When the type being created is a record with default initializers, the return
value is void but that's not what we want to use to determine the type of the
argument.  So, when handling resolveNew for a concrete record, if we know we're
the last statement of a block for an ArgSymbol, add a SymExpr to the temporary
we send in to the initializer, so that we can get the right type back.

This is a hack and I am not proud of it, but I'm not sure there's a better way,
aside from redoing argument handling.

Also, currently only applies to concrete records.  Need to expand to generic
records and generic record instantiations, but I'm going to test this version
first as a proof of concept.