Add barrier subtimings to isx-hand-optimized
Previously the exchange subtimer timed the exchange as well as the barrier.
This adds an extra subtimer to the barrier step, to hopefully allow us to see
if the exchange step is completely in roughly the same amount of time, or if
some tasks are spending more time just waiting for the barrier.