Name generated executable after main module's filename
Implements the proposal in issue #8474.

In issue #6283 / PR #6811, we changed the default executable name
created by the compiler from a.out to the name of the main
module. Every Chapel program defines a unique main module so this
results in an unambiguous name, and often the main module takes its
name from the file containing it, so in those cases, this results in
chpl testit.chpl creating an executable named testit.

However, when the main module and the filename don't match, this can
cause significant confusion, which has led to the following
counter-proposal: Perhaps we should be naming the executable after the
file containing the main module rather than the main module itself.