Default to using CHPL_TASKS=fifo for freebsd
qthreads doesn't currently build on netbsd or freebsd. We defaulted to fifo for
netbsd with https://github.com/chapel-lang/chapel/pull/1244, this just does the
same for freebsd.

With this, `vagrant/chapeldefaultcheck.sh` passes on my mac and an ubuntu 17.10
machine. Without it, the freebsd builds would fail when building qthreads.