Fix auto-destroy overloads for Owned with coercion to class
The overload resolution rules indicate that a generic with
any type (i.e. not restricted at all) is always less specific
than another argument, even if the other argument requires coercion.

In this case, that means that the autoDestroy(x:object) was
being preferred to autoDestroy(x) for something of type Owned,
which would mean they were never destroyed and the object freed.

The resolution is to make minor adjustments to overloads in ChapelBase
that are in this situation, so that the :object case and the generic
case are handled with similar argument types (namely both are now
handled with fully generic arguments).