Reduce contention from ugni's polling thread in isx-hand-optimized
This just enables the opt-in mechanism added in #8562 for isx-hand-optimized

For ISx we ideally want all tasks to be operating in lockstep so that they all
hit the exchange barrier at the same time so we don't waste time waiting for
tasks to catch up. However, the progress thread creates contention with tasks
running on the same core, so those tasks are noticeably slower even for parts
of the computation that are completely local. Prior to this change for the
input stage (which just creates an array and fills it with random values) we'd
see timings like:

    input = 0.9 avg (0.81 min .. 1.28 max)

where the max value is from the task sharing a core with the progress thread.
With this change we see much tighter timings:

    input = 0.89 avg (0.81 min .. 0.95 max)

I think the remaining variability in timings is because of the serialized
component of dynamic array registration. We have other work to reduce the
critical section for array registration (#8177), but prior to this change (and
the scalable barrier) we had a really hard time evaluating the impact of that
work since timings were so variable to begin with.