Convert chpl__initCopy/autoCopy over to copy initializers for ZMQ
Under valgrind, once we upgraded to SLES12, one of the ZMQ tests
(server.chpl) started getting failures which Preston traced back to
the point when initializers were introduced for this module.  Checking
the code quickly, I realized that I had failed to replace the
chpl__initCopy/autoCopy functions with copy initializers, so have now
done that and the valgrind error has gone away.

In last night's testing, Greg reported that the gasnet
configurations were timing out for the same test, so hopefully this
will resolve that issue as well.