Do not flag "weakly"-recursive iterators
find_recursive_caller() in lowerIterators.cpp used to cause FLAG_RECURSIVE_ITERATOR on iterators that are recursive through non-iterator calls. For example, i1() and i2() got this flag in the following example:

  iter i1() {
    for i2() do ...;
  }

  iter i2 {
    foo();
    yield ...;
  }

  proc foo() {
    for i1() do ...;
  }

For the purposes of lowerIterators, however, this is not necessary. For example of the above example, i1() and i2() can be inlined into foo(). This will make foo() call itself recursively. However, it will not require us to resort to recursive-iterator implementation machinery.

This change eliminates FLAG_RECURSIVE_ITERATOR in cases like this, allowing us to use non-recursive-iterator implementation in more cases. Which hopefully simplifies the generated code.