Add perfstats to measure dynamic alloc related activities.
This adds some more internal performance statistics, this time having to
do with dynamic allocation.  I also did some cleaning up in perfstats,
reducing the number of times we repeat the counter type and things like
that.  All of this is normally turned off; we only use it for low-level
performance investigations because it does slow down normal operations
noticeably.