Improve start_test tab completion by hiding "script oriented" flags
Hide some flags that are really meant to be used by other scripts, and aren't
helpful for developers. This is just meant to make tab completion more useful
by restricting the suggestions to flags developers will actually use instead of
polluting the suggestions with useless ones. Note that they can still be
printed with --help-all if you really want to see them all.

This hides:
 --compiler
 --performance-description
 --performance-configurations
 --compperformance
 --gen-graphs
 --nodisplaygraphrange
 --graphsgendefault
 --gengraphopts
 --nostdinredirect
 --stdinredirect
 --launchertimeout
 --no-chpl-home-warn
 --progress
 --test-root
 --junit-xml
 --junit-xml-file
 --junit-remove-prefix