Update references to generated executable filename
This commit updates files that mention the generated executable's name
to reflect the revised behavior.  I found these locations by going
through those that Michael changed in PR #6978 when he originally
changed the executable filename away from 'a.out'.  I also fixed a
typo that was introduced in that PR.