Enforce phase-one by default for records
By enforcing phase-one by default for records, we can start to do away
with the old initializer proposal and just focus on the new proposal
semantics.

A method, 'completePhase0', has been added to InitNormalize to represent
a transition upon encountering a super.init().

InitNormalize::startPhase has been updated to choose phase one by
default for records.

Note that with this commit we are still in a hybrid world for records.
Notably we still need to disallow super.init inside a record. This error
has been left out for now to reduce the diff of the PR containing this
commit.