Fix locale model init phase violations (#8690)
Fix locale model init phase violations (#8690)

[reviewed by @benharsh , with @mlchu being asked for post-commit review of APU locale model]

The locale models were updated to use initializers while the rules were new and in flux. #8660 started enforcing some of the rules (in particular, not passing "this" to something during phase 1), so the locale models need to be updated to correct that problem. While doing this, I noticed other violations of the rules that are not yet being enforced. Currently, const members should not be initialized in phase 2, but they are.

This change gets testing back on track by fixing the enforced violations, and fixes violations anticipated to be enforced in the future.

In some places, a const field had to be changed to a var field to avoid a full refactoring of locale models. In the long run, all locale models will need to be refactored to take new init rules into account.

 Passed hellos on the knl locale model
 Passed release/examples on the numa locale model
 Passed full local testing on the flat locale model