Restore duplicate copy of versionButton.php
This PR restores the duplicate copy of versionButton.php that I was
trying to avoid because when building RPMs, we get a complaint for
the symbolic link to a nonexistent file.  This is designed to fix
the build, but still leaves something to be desired:

* we're back to having a duplicate file
* the versionButton isn't actually particularly useful outside of
  the web context (within the RPM) since there aren't multiple
  versions available in the RPM.

I'm going to think more about a longer-term solution after the
build is back to working again.