Add five test records and initializers with arrays
Default initializers revealed an issue when a record stored as a field an array
of other records that used default initializers.  These tests demonstrate that
the problem only occurs when both types utilize default initializers.

However, there is a separate issue with inferring the type for the explicit
initializer array argument (which may be due to having written it wrong, idk)