Fix default binary name for foo/bar/baz.chpl
Paul Cassella pointed out that with my recent change to the default
executable names, I accidentally made it so that foo/bar/baz.chpl
would default to foo/bar/baz rather than simply ./baz.  This fixes
that by searching backwards for the last slash and starting the
executable name from the character following, if it exists.