Make study version avoid enum->int coercions
I'm exploring the question of what the impact would be of disabling
enum->int coercions.  Fasta relies on this at present and there are
a few ways of rewriting it, some of which would be more onerous,
others less so.  This is one of the less onerous ways, but one that
may impact performance adversely.  Updating my study version to
see what the degree of impact is in testing.