Experimental support for the Allinea compiler, ARM's commercial compiler (#8742)
Experimental support for the Allinea compiler, ARM's commercial compiler (#8742)

[reviewed by @bradcray ]

On Cray XC50-AC systems (64-bit ARM), the PrgEnv-intel compiler is replaced with PrgEnv-allinea, which is ARM's commercial compiler.

I added this experimental support to see if we should include PrgEnv-allinea in the XC50-AC version of the Chapel 1.17 module build. As it turns out, configuration of that compiler on XC50-AC systems is not complete, so this Chapel support is not operational yet. However, I would like to preserve what I have done so far, so we can use it when that compiler comes on-line.

Documentation will wait until we are actually able to claim normal support for the Allinea compiler.