Remove more unused code.
The "with (ref x0)" clauses on the coforall-stmts across worker nodes
and cores were only needed when the references to that variable were
directly in the bodies of those stmts.  Now that they're in a separate
procedure, even though it's inlined, they clearly refer to the global
and we don't need those clauses.