Replace the homegrown barrier with a standard allLocalesBarrier.
The barrier improvements done in #8550 and follow-ons hadn't quite
arrived when this test was written, so it had a homegrown barrier to
get the many-to-one load from all the initiating tasks aligned in
time.  Here, switch to the standard allLocalesBarrier.  This change
is largely the work of @ronawho.

In testing this seems to have improved execution time stability and the
rates we measure when doing GETs and fetching AMOs.  Those used to be
rather variable.  So the original homegrown barrier probably had some
unappreciated problems.